@Samara-Alves [Task:Atualizando camadas do blog pessoal]

+++ src/main/java/org/generation/blogPessoal/controller/PostagemController.java
@@ -1,6 +1,9 @@
 package org.generation.blogPessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.generation.blogPessoal.model.Postagem;
 import org.generation.blogPessoal.repository.PostagemRepository;
@@ -15,47 +18,56 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 @RequestMapping("/postagens")
-@CrossOrigin("*")
+@CrossOrigin(value = "*", allowedHeaders = "*")
 public class PostagemController {
-	 
+
 	@Autowired
 	private PostagemRepository repository;
-	
-	@GetMapping
-	public ResponseEntity<List<Postagem>>GetAll(){
-		return ResponseEntity.ok(repository.findAll());
-		
-	}
-    @GetMapping("/{id}")
-    public ResponseEntity<Postagem> GetById(@PathVariable long id){
-    	return repository.findById(id)
-    			.map(resp -> ResponseEntity.ok(resp))
-    			.orElse(ResponseEntity.notFound().build());
-    	
-    }
-    @GetMapping("/título/{título}")
-    public ResponseEntity<List<Postagem>> GetByTítulo(@PathVariable String título) {
-    	return ResponseEntity.ok(repository.findAllByTituloContainingIgnoreCase(título));
-    	
-    }
-    @PostMapping 
-    public ResponseEntity<Postagem> post (@RequestBody Postagem postagem){
-    	return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(postagem));
-    	
-    	
-    }
-    @PutMapping 
-    public ResponseEntity<Postagem> put (@RequestBody Postagem postagem){
-    	return ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem));
-    	}
-    
-       @DeleteMapping("/{id}")
-       public void delete(@PathVariable long id) {
-    	   repository.deleteById(id);
-       }
-}
 
+	@GetMapping
+	public ResponseEntity<List<Postagem>> getAll() {
+		return ResponseEntity.ok(repository.findAll());
+
+	}
+
+	@GetMapping("/{id}")
+	public ResponseEntity<Postagem> getById(@PathVariable long id) {
+		return repository.findById(id).map(resp -> ResponseEntity.ok(resp)).orElse(ResponseEntity.notFound().build());
+
+	}
+
+	@GetMapping("/título/{título}")
+	public ResponseEntity<List<Postagem>> GetByTítulo(@PathVariable String título) {
+		return ResponseEntity.ok(repository.findAllByTituloContainingIgnoreCase(título));
+
+	}
+
+	@PostMapping
+	public ResponseEntity<Postagem> post(@Valid @RequestBody Postagem postagem) {
+		return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(postagem));
+
+	}
+
+	@PutMapping
+	public ResponseEntity<Postagem> put(@Valid @RequestBody Postagem postagem) {
+		return repository.findById(postagem.getId())
+				.map(resp -> ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem)))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
+	}
+
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+	@DeleteMapping("/{id}")
+	public void delete(@PathVariable long id) {
+		Optional<Postagem> post = repository.findById(id);
+		
+		if(post.isEmpty())
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+		repository.deleteById(id);
+	}
+}
