+++ src/main/java/org/generation/blogPessoal/controller/TemaController.java
@@ -1,6 +1,9 @@
 package org.generation.blogPessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.generation.blogPessoal.model.Tema;
 import org.generation.blogPessoal.repository.TemaRepository;
@@ -15,11 +18,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 @CrossOrigin (origins = "*", allowedHeaders = "*")
-@RequestMapping("/tema")
+@RequestMapping("/temas")
 public class TemaController {
 	
 	@Autowired
@@ -31,31 +36,37 @@ public class TemaController {
 		
 	}
 	@GetMapping("/{id}")
-	public ResponseEntity<Tema> getById(@PathVariable(value = "id")long id){
+	public ResponseEntity<Tema> getById(@PathVariable long id){
 		return repository.findById(id).map(resp -> ResponseEntity.ok(resp))
-				.orElse(ResponseEntity.notFound().build());
-		
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());	
 	}
 	
-	@GetMapping("/nome{nome}")
-	public ResponseEntity<List<Tema>> getByName(@PathVariable String nome) {
-		return ResponseEntity.ok(repository.findAllByDescricaoContainingIgnoreCase(nome));
+	@GetMapping("/descricao{descricao}")
+	public ResponseEntity<List<Tema>> getByTitulo(@PathVariable String descricao) {
+		return ResponseEntity.ok(repository.findAllByDescricaoContainingIgnoreCase(descricao));
 		
 	}
     @PostMapping
-    public ResponseEntity<Tema> post (@RequestBody Tema tema){
+    public ResponseEntity<Tema> post (@Valid @RequestBody Tema tema){
     	return ResponseEntity.status(HttpStatus.CREATED)
     			.body(repository.save(tema));
     }
     @PutMapping
-    public ResponseEntity<Tema> put (@RequestBody Tema tema){
-    	return ResponseEntity.status(HttpStatus.OK).body(repository.save(tema));   	
+    public ResponseEntity<Tema> put (@Valid @RequestBody Tema tema){
+    	return repository.findById(tema.getId())
+				.map(resp -> ResponseEntity.status(HttpStatus.CREATED).body(repository.save(tema)))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());  	
      			
     }
- 
+    @ResponseStatus
     @DeleteMapping("/{id}")
     public void delete(@PathVariable long id) {
-    	repository.deleteById(id);
+    	Optional<Tema> tema = repository.findById(id);
+
+		if(tema.isEmpty())
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
+		    	repository.deleteById(id);
     }
 }
